deploy/requirements.txt
boto3>=1.35.0
botocore>=1.35.0
oci>=2.150.0
python-dotenv>=1.0.1
